정하영

+++ src/main/java/com/sports/user/UserService.java
@@ -44,7 +44,7 @@ public class UserService {
     }
 
     public User findById(String userId) {
-        return userRepository.findById(Integer.valueOf(userId))
+        return userRepository.findById(Long.valueOf(userId))
                 .orElseThrow(() -> new RuntimeException("유저를 찾을 수 없습니다.")); // 예외 처리
     }
